requirements-dev.txt
alabaster==0.7.12
attrs==21.4.0
Babel==2.9.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
cryptography==36.0.1
docutils==0.17.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.1
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.0.3
keyring==23.5.0
livereload==2.6.3
MarkupSafe==2.0.1
multidict==6.0.2
oauthlib==3.2.0
packaging==21.3
pkginfo==1.8.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pytest==6.2.5
pytest-order==1.0.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
readme-renderer==32.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tornado==6.1
tqdm==4.62.3
twine==3.8.0
urllib3==1.26.8
vcrpy==4.1.1
webencodings==0.5.1
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
